Moving handler wrappers back out
